+++ flutter_custom_tabs/test/mocks/mock_custom_tabs_platform.dart
@@ -50,9 +50,7 @@ class MockCustomTabsPlatform extends Fake
     returnSession = prewarmingSession;
   }
 
-  void setInvalidateExpectations({
-    required PlatformSession session,
-  }) {
+  void setInvalidateExpectations({required PlatformSession session}) {
     argSession = session;
   }