+++ flutter_custom_tabs_android/test/types/custom_tabs_animations_test.dart
@@ -4,12 +4,14 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   group('CustomTabsAnimations', () {
-    test('toMessage() returns empty message when animation values are null',
-        () {
-      const animations = CustomTabsAnimations();
-      final actual = animations.toMessage();
-      expect(actual, isEmpty);
-    });
+    test(
+      'toMessage() returns empty message when animation values are null',
+      () {
+        const animations = CustomTabsAnimations();
+        final actual = animations.toMessage();
+        expect(actual, isEmpty);
+      },
+    );
 
     test('toMessage() returns a message with complete options', () {
       const animations = CustomTabsAnimations(
@@ -31,15 +33,12 @@ void main() {
   group('CustomTabsSystemAnimations', () {
     test('slideIn: creates the built-in slide in animation', () {
       final actual = CustomTabsSystemAnimations.slideIn().toMessage();
-      expect(
-        actual,
-        <String, String>{
-          'startEnter': 'android:anim/slide_in_right',
-          'startExit': 'android:anim/slide_out_left',
-          'endEnter': 'android:anim/slide_in_left',
-          'endExit': 'android:anim/slide_out_right',
-        },
-      );
+      expect(actual, <String, String>{
+        'startEnter': 'android:anim/slide_in_right',
+        'startExit': 'android:anim/slide_out_left',
+        'endEnter': 'android:anim/slide_in_left',
+        'endExit': 'android:anim/slide_out_right',
+      });
     });
 
     test('fade: creates the built-in fade animation', () {