+++ flutter_custom_tabs_android/test/types/partial_custom_tabs_test.dart
@@ -195,17 +195,15 @@ void main() {
     test('returns associated value', () {
       expect(
         CustomTabsActivitySideSheetRoundedCornersPosition
-            .defaultPosition.rawValue,
+            .defaultPosition
+            .rawValue,
         0,
       );
       expect(
         CustomTabsActivitySideSheetRoundedCornersPosition.none.rawValue,
         1,
       );
-      expect(
-        CustomTabsActivitySideSheetRoundedCornersPosition.top.rawValue,
-        2,
-      );
+      expect(CustomTabsActivitySideSheetRoundedCornersPosition.top.rawValue, 2);
     });
   });
 }