+++ flutter_custom_tabs/CHANGELOG.md
@@ -1,3 +1,14 @@
+## NEXT
+
+- Updates the minimum supported SDK version to Flutter 3.44/Dart 3.12.
+- Migrates to built-in Kotlin.
+- Updates minimum supported Kotlin version to 2.2.10.
+- Migrates the build configuration from Groovy to Kotlin in the android example app.
+- Updates the minimum supported SDK version to Android SDK 23.
+- Updates minimum supported version to iOS 13.
+- Adopts `UISceneDelegate` in the iOS example app.
+- Updates minimum required `flutter_lints` version to 6.0.0.
+
 ## 2.5.0
 
 - Updates minimum supported SDK version to Flutter 3.27.0/Dart 3.6.