+++ flutter_custom_tabs_platform_interface/CHANGELOG.md
@@ -1,8 +1,13 @@
+## NEXT
+
+- Updates the minimum supported SDK version to Flutter 3.44/Dart 3.12.
+- Updates minimum required `flutter_lints` version to 6.0.0.
+
 ## 2.4.0
 
 - Updates minimum supported SDK version to Flutter 3.27.0/Dart 3.6.
-- Updates minimum required mockito version to 5.4.4.
-- Updates minimum required flutter_lints version to 5.0.0.
+- Updates minimum required `mockito` version to 5.4.4.
+- Updates minimum required `flutter_lints` version to 5.0.0.
 
 ## 2.3.0